% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotropy_blockpermutation.R
\name{isotropy_blockpermutation}
\alias{isotropy_blockpermutation}
\title{Isotropy test based on the variogram using block permutation}
\usage{
isotropy_blockpermutation(
  data,
  lagmat = rbind(c(1, 0), c(0, 1), c(1, 1), c(1, -1)),
  A = rbind(c(1, -1, 0, 0), c(0, 0, 1, -1)),
  estimator = "all",
  window.dims = c(6, 6),
  var.robust = TRUE,
  c.regularisation = 0.1,
  B = 1000
)
}
\arguments{
\item{data}{A data frame with three column containing the data (first column) and the grid (second and third column).}

\item{lagmat}{A matrix of spatial lags, where each row represents a lag vector specified as \eqn{(lag.x, lag.y)} for which the semivariogram value is to be estimated.}

\item{A}{A contrast matrix, where each row defines a contrast of the estimated semivariogram evaluated at the lags specified in \code{lagmat}.}

\item{estimator}{A character string or vector of character strings specifying the estimators to be used. Possible values are \code{"matheron"}, \code{"genton"}, \code{"mcd"}. For \code{"all"} all three estimators are used. See details. Default is \code{"all"}.}

\item{window.dims}{A length-two vector specifying to the width and height of the subblocks -given as the number of columns and rows, respectively- used in the block permutation approach. See details.}

\item{var.robust}{Logical. If \code{TRUE} a robust covariance estimator is used for estimation of the covariance matrix, otherwise the classical sample covariance matrix estimator is used. See details. Default is \code{TRUE}.}

\item{c.regularisation}{Numeric. A value between 0 and 1. Needed for the regularisation of the covariance estimation, if this is singular. See Gierse and Fried (2026); default: 0.1.}

\item{B}{Numeric. Number of permutations in the block permutations approach. Default is 1000.}
}
\value{
An object of class "isotropyRob" which is a list with an list for each estimator containing the following elements
\itemize{
\item teststatistic: value of the test statistic
\item p.value: calculated p-value
\item cov: a list with the following two elements:
\itemize{
\item cov: a matrix containing the estimated covariance matrix of the estimated semivariogram vector for all lags of \code{lagmat}
\item gamma.sub: a matrix with the semivariogram estimations for each permutated data set
}
\item regularization: information if the regularized version of the covariance estimation is used in the test statistic}
}
\description{
Returns robust and non-robust isotropy test based on the variogram using a block permutation approach for test decision
}
\details{
This functions implements the block permutation test for isotropy proposed by Gierse and Fried (2026).  The test is based on the test idea of Guan et al. (2004).

The idea of the test is to compare variogram estimations for lag vectors with the same distance but different directions using a contrast test. A block permutation approach
with small non-overlapping blocks is used for estimation of the covariance matrix of the variogram estimation and for the p-value calculation. For more details see  Gierse and Fried (2026).

Implemented are three different estimators the Matheron variogram estimator (Matheron, 1962), the robust Genton estimator (Genton, 1998) and the robust MCD.diff estimator (Gierse and Fried, 2025).
More information about the implementation can be found in documentations of the functions
\code{\link{variogram_est}} and \code{\link{variogram_est_general}}.
}
\examples{
\donttest{
 ## Simulation of an anisotrop GRF with an quadratic outlier block
 dat <-  simulate_grf(gridsize = c(24, 24), param.variogram = c(1, 6),
                      aniso.param = c(pi/4, 2), out.type = "block",
                      block.type = "square", amount = 0.1,
                      param.outlier = c(0,5), n.it = 1)

 dat <- cbind(dat$data, dat$grid)

 ## Test for isotropy using 6x6 blocks
 test <- isotropy_blockpermutation(dat, window.dims = c(6,6))
}

}
\references{
Genton, M. (1998). Highly robust variogram estimation. \emph{Mathematical Geology}, 30, 213-221. \doi{https://doi.org/10.1023/A:1021728614555}

Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}

Gierse, J. & Fried. R. (2026). A robust nonparametric test for spatial isotropy in lattice data, \emph{arXiv}. \doi{https://doi.org/10.48550/arXiv.2605.18030}

Guan, Y., Sherman, M. & Calvin, J. A. (2004). A nonparametric test for spatial isotropy using subsampling. \emph{J. Americ. Statist. Assoc.}, 99, 810-821. \doi{https://doi.org/10.1198/016214504000001150}

Matheron, G. (1962). Traité de géostatistique appliquée, Tome I. \emph{Mémoires du Bureau de Recherches Géologiques et Miniéres}, no. 14, Editions Technip, 333
}
\seealso{
\code{\link{isotropy_test}}, \code{\link{isotropy_subsampling}}
}
