% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotropy_subsampling.R
\name{isotropy_subsampling}
\alias{isotropy_subsampling}
\title{Test of isotropy based on the variogram using subsampling techniques}
\usage{
isotropy_subsampling(
  data,
  lagmat = rbind(c(1, 0), c(0, 1), c(1, 1), c(1, -1)),
  A = rbind(c(1, -1, 0, 0), c(0, 0, 1, -1)),
  estimator = "all",
  window.dims,
  var.robust = TRUE,
  edge = TRUE,
  c.regularisation = 0.1
)
}
\arguments{
\item{data}{A dataframe with three columns containing the data (first column) and the grid (second and third column).}

\item{lagmat}{A matrix of spatial lags, where each row represents a lag vector specified as \eqn{(lag.x, lag.y)} for which the semivariogram value is to be estimated.}

\item{A}{A contrast matrix, where each row defines a contrast of the estimated semivariogram evaluated at the lags specified in \code{lagmat}.}

\item{estimator}{A character string or vector of character strings specifying the estimators to be used. Possible values are \code{"matheron"}, \code{"genton"}, \code{"mcd"}. If \code{"all"} all three estimators are used. See details. Default is \code{"all"}.}

\item{window.dims}{A length-two vector specifying to the width and height of the moving windows -given as the number of columns and rows, respectively- used in the subsampling approach.}

\item{var.robust}{Logical. If \code{TRUE} a robust covariance estimator is used for estimation of the covariance matrix else the classical sample covariance matrix estimator is used. See details. Default is \code{TRUE}.}

\item{edge}{Logical. If \code{TRUE} a correction for edge effects is used in the supsampling approach as suggested in Guan et. al. (2004). Default is \code{TRUE}.}

\item{c.regularisation}{Numeric. A value between 0 and 1. Needed for the regularisation of the covariance estimation, if this is singular. See Gierse and Fried (2026); default: 0.1.}
}
\value{
An object of class "isotropyRob" with a list for each estimator containing the following elements
\itemize{
\item teststatistic: value of the test statistic
\item p.value: calculated p-value
\item cov: a list with the following two elements:
\itemize{
\item cov: a matrix containing the estimated covariance matrix of the estimated semivariogram vector for all lags of \code{lagmat}
\item gamma.sub: a matrix with the semivariogram estimations for each subsample
}
\item regularisation: information if the regularised version of the covariance estimation is used in the test statistic}
}
\description{
Returns the nonparametric test of isotropy from Guan et. al. (2004) based on the variogram using a subsampling technique for test decision for data on a regular grid. See Guan et. al. (2004) for more details.
}
\details{
This functions implements the subsampling test for isotropy of Guan et al. (2004) as well as a version of the test, where the Matheron variogram estimator is
replaced by robust variogram estimators. For more information about this see Gierse and Fried (2026).

The idea of the test is to compare variogram estimations for lag vectors with the same distance but different directions using a contrast test. A subsampling approach
with small overlapping subsamples is used for estimation of the covariance matrix of the variogram estimation and for the p-value calculation.

Implemented are three different estimators: the Matheron variogram estimator (Matheron, 1962), the robust Genton estimator (Genton, 1998) and the robust MCD.diff estimator (Gierse and Fried, 2025).
More information about the implementation can be found in the documentations of the functions
\code{\link{variogram_est}} and \code{\link{variogram_est_general}}.
}
\examples{
\donttest{
 ## Simulation of an anisotrop GRF with an quadratic outlier block
 dat <-  simulate_grf(gridsize = c(24, 24), param.variogram = c(1, 6),
                      aniso.param = c(pi/4, 2), out.type = "block",
                      block.type = "square", amount = 0.1,
                      param.outlier = c(0,5), n.it = 1)

 dat <- cbind(dat$data, dat$grid)

 ## Test for isotropy using 5x5 subsamples
 test <- isotropy_subsampling(dat, window.dims = c(5,5))
}

}
\references{
Genton, M. (1998). Highly robust variogram estimation. \emph{Mathematical Geology}, 30, 213-221. \doi{https://doi.org/10.1023/A:1021728614555}

Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}

Gierse, J. & Fried. R. (2026). A robust nonparametric test for spatial isotropy in lattice data, \emph{arXiv}. \doi{https://doi.org/10.48550/arXiv.2605.18030}

Guan, Y., Sherman, M. & Calvin, J. A. (2004). A nonparametric test for spatial isotropy using subsampling. \emph{J. Americ. Statist. Assoc.}, 99, 810-821. \doi{https://doi.org/10.1198/016214504000001150}

Matheron, G. (1962). Traité de géostatistique appliquée, Tome I. \emph{Mémoires du Bureau de Recherches Géologiques et Miniéres}, no. 14, Editions Technip, 333
}
\seealso{
\code{\link{isotropy_test}}, \code{\link{isotropy_blockpermutation}}
}
