% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_correctionfactors.R
\name{simulate_correctionfactors}
\alias{simulate_correctionfactors}
\title{Simulate of correctionfactors for the variogram estimators for small grids}
\usage{
simulate_correctionfactors(
  gridsize,
  variogram.est,
  n.it = 1000,
  estimator,
  reweighting = TRUE,
  variogram = "Sph",
  range = NULL,
  sill = NULL,
  nugget = NULL,
  seed = NULL
)
}
\arguments{
\item{gridsize}{A two-dimensional vector containing the gridsize in x-direction (nx) and the gridsize in y-direction (ny).}

\item{variogram.est}{An object of class \code{varioRob} or \code{varioRob}, i.e. an output of the functions \code{\link{variogram_est}} or \code{\link{variogram_est_general}}.}

\item{n.it}{Numeric. Number of iterations used for the simulation of the correctionfactors; default: 1000.}

\item{estimator}{A character string or vector of character strings specifying the estimators to be used. Possible estimators are \code{"matheron"}, \code{"genton"}, \code{"mcd.org"} or \code{"mcd.diff"}. If \code{"all"} all four estimators are used. See details; Default is \code{"all"}.}

\item{reweighting}{Logical indicating if the reweighted version of the MCD based variogram estimators should be used; default is \code{TRUE}.}

\item{variogram}{The variogram model. See \code{vgm} of the package \pkg{gstat} for more information; default \code{"Sph"}.}

\item{range}{Numeric. A start value for the range estimation. If \code{NULL} the start value is the half of the largest lag; default is \code{NULL}.}

\item{sill}{Numeric. A start value for the sill estimation. If \code{NULL} the start value is the mean of the variogram estimation for the three highest lags; default is \code{NULL}.}

\item{nugget}{Numeric. A start value for the nugget estimation. If \code{NULL} the start value is the 0; default is \code{NULL}.}

\item{seed}{Numeric. A possible seed for randomnumber generation.}
}
\value{
A list with two elements
\itemize{
\item correction factors: a matrix with the simulated correction factors with a column for each estimator and a row for each direction
\item ranges: a vector with the estimated range for each direction
}
}
\description{
Returns simulated finite sample correction factors for (non-)robust directional variogram estimators using a gaussian random field.
}
\details{
For small grid sizes and if many lags in one direction should be estimated, the robust estimator needs multiplicative finite sample correction factors (see Gierse & Fried, 2025).
The correction factors are simulated based on the grid size as well as the lags of interest using gaussian random fields. The procedure is the following:

1. estimate the range of the variogram given the variogram estimation for the in \code{variogram} specified variogram model using a least square estimator

2. simulate data (gaussian random field) for the variogram model using the estimated range, a sill of 1 and a nugget of 0

3. estimate the variogram for all simulated data sets

4. calculate the correction factors (see Gierse & Fried, 2025)

Only the estimated range is used for simulation of the data, since the sill as well as the nugget effect cannot be estimated from the uncorrected data.

For the simulation of the data the function \code{predict} of the package \pkg{gstat} is used.

If no start values for the variogram parameters are defined, they are specified as follows
\itemize{
    \item range: half of the largest estimated lag
    \item sill: mean of the last three variogram estimates
    \item nugget: 0
 }

More information about the implementation can be found in documentations of the functions
\code{\link{variogram_est}} and \code{\link{variogram_est_general}}.
}
\examples{
\donttest{
## Simulate an isotropic gaussian random field without outliers
dat <- simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6), n.it = 1)

dat <- cbind(dat$data, dat$grid)

## calculate the variogram using different robust estimators for all four directions
varog <- variogram_est(data = dat, hmax = c(7,7,5,5),
                       direction = c("S-N", "E-W", "SW-NE", "SE-NW"),
                       estimator = c("genton", "mcd.diff"))

## simulate correction factors
corr <- simulate_correctionfactors(gridsize = c(20, 20),
                                   variogram.est = varog,
                                   n.it = 1000,
                                   estimator = c("genton", "mcd.diff"))
}

}
\references{
Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}
}
