% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_grf.R
\name{simulate_grf}
\alias{simulate_grf}
\title{Simulation of gaussian random fields on a regular grid with different types of outliers}
\usage{
simulate_grf(
  gridsize,
  dist.outlier = rnorm,
  variogram = "spherical",
  param.variogram,
  nugget = NULL,
  aniso.param = NULL,
  out.type = NULL,
  block.type = NULL,
  amount = NULL,
  param.outlier = NULL,
  mixture = TRUE,
  n.it = 1000,
  seed = NULL
)
}
\arguments{
\item{gridsize}{Integer vector of length 2 containing the gridsize in x-direction and the gridsize in y-direction}

\item{dist.outlier}{A function to draw random numbers from the distribution of the outliers; default: \code{rnorm}}

\item{variogram}{The variogram model. See function \code{cov.spatial} of the package \pkg{geoR} for more information; default: \code{spherical}.}

\item{param.variogram}{A two-dimensional vector containing the parameters of the variogram. The first element specifies the (partial) sill and the second one the range of the variogram.}

\item{nugget}{Numeric. Specifies the nugget effect; default: \code{NULL}}

\item{aniso.param}{A two-dimensional vector containing the parameters of the geometric anisotropy. The first element is the rotation angle and the second one the anisotropy ratio. If \code{NULL} an isotropic random field is simulated; default: \code{NULL}.}

\item{out.type}{A character string specifying the type of outlier (\code{"block"} or \code{"isolated"}. If \code{NULL} a random field without outliers is simulated. See also details; default: \code{NULL}.}

\item{block.type}{A character string only needed for generation of block outliers. It specifies wheather the outlier block should be nearly quadratic (\code{"square"}), as rectangular as possible (\code{"rectangle"}) or if it is random (\code{"random"}).}

\item{amount}{Numeric. A value between 0 and 1 specifying the amount of outlier; default: \code{NULL}.}

\item{param.outlier}{A vector specifying the parameters of the outlier distribution; default: \code{NULL}.}

\item{mixture}{logical. Specifying whether the outlier should be generated as mixture model (\code{TRUE}) or as additive outliers (\code{FALSE}); default \code{TRUE}.}

\item{n.it}{Numeric. Number of iterations; default: \code{1000}.}

\item{seed}{Numeric. A possible seed for randomnumber generation.}
}
\value{
A list with two elements
\itemize{
\item data: a matrix with n.it columns; in each column is one simulated data set
\item grid: a data.frame with two columns; the first column contains the x-coordinate and the second column the y-coordinate
}
}
\description{
Returns simulated (an)isotropy gaussian random fields on a regular grid without outliers, with isolated outliers, or with an outlier block.
}
\details{
Gaussian random fields (GRF) without outliers, with isolated outliers or with block outliers can be simulated. The GRF without outliers is simulated
         using the function \code{grf} of package \pkg{geoR}.
         GRFs with outliers are generated in the following manner:

         Isolated outliers (occur randomly of the grid):
         \eqn{Y(\boldsymbol{s})|N_{n_0} = \begin{cases} Z(\boldsymbol{s}),~\text{ if } \boldsymbol{s} \notin N_{n_0} \\ W,~ \text{ if } \boldsymbol{s} \in N_{n_0}\end{cases}}

         Block outliers (occur spatially aggregated in a block):
         \eqn{Y(\boldsymbol{s})|(U=\boldsymbol{s}_0) = \begin{cases} Z(\boldsymbol{s}),~\text{ if } \boldsymbol{s} \notin N(\boldsymbol{s}_0) \\W,~ \text{ if } \boldsymbol{s} \in N(\boldsymbol{s}_0)\end{cases}}

         Thereby \eqn{Z(\boldsymbol{s})} is the GRF without outliers, \eqn{N_{n_0}} is a set of \eqn{n_0} randomly selected locations of the grid, \eqn{W} is the distribution of the outliers, \eqn{N(\boldsymbol{s}_0)} is a block neighbourhood around the gridpoint \eqn{\boldsymbol{s}_0} and \eqn{U} is a random variable which select one gridpoint.
         For more information see Gierse and Fried (2025).
}
\examples{
\donttest{
 ## Simulation of an anisotrop GRF without outliers
 dat1 <-  simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6),
                       aniso.param = c(pi/4, 2), n.it = 1)

 ## Simulation of an anisotrop GRF with isolated outliers
 dat2 <-  simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6),
                       aniso.param = c(pi/4, 2), out.type = "isolated",
                       amount = 0.1, param.outlier = c(0,5), n.it = 1)

 ## Simulation of an anisotrop GRF with an quadratic outlier block
 dat2 <-  simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6),
                       aniso.param = c(pi/4, 2), out.type = "block",
                       block.type = "square", amount = 0.1,
                       param.outlier = c(0,5), n.it = 1)

}

}
\references{
Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}
}
