% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_est.R
\name{variogram_est}
\alias{variogram_est}
\title{Nonparametric directional variogram estimation}
\usage{
variogram_est(
  data,
  hmax,
  direction = c("S-N", "E-W", "SW-NE", "SE-NW"),
  estimator = "all",
  cp = NULL,
  reweighting = TRUE,
  ...
)
}
\arguments{
\item{data}{A dataframe with three columns containing the data (first column) and the grid (second and third column).}

\item{hmax}{A vector with the maximal lag to be estimated for each entry in \code{direction}. If it has length 1 the same hmax is used for all directions.}

\item{direction}{A vector containing the directions to be estimated. Possible values are \code{"E-W"} (east-west), \code{"S-N"} (south-north), \code{"SW-NE"} (southwest-northeast) and \code{"SE-NW"} (southeast-northwest). Default is all four directions.}

\item{estimator}{A character string or vector of character strings specifying the estimators to be used. Possible estimators are \code{"matheron"}, \code{"genton"}, \code{"mcd.org"} or \code{"mcd.diff"}. If \code{"all"} all four estimators are used. See details; Default is "all".}

\item{cp}{A vector containing possible finite sample correction factors for each direction for the MCD based variogram estimators. If it has length 1 the same correction factor is used for all directions.}

\item{reweighting}{Logical indicating if the reweighted version of the MCD based variogram estimators should be use. Default is \code{TRUE}.}

\item{...}{Optional argument for the function \code{covMcd} of the package \pkg{robustbase} used for the MCD based variogram estimators}
}
\value{
An object of class "varioRob" which is a list with an dataframe for each direction with columns
\itemize{
\item lag.x: lag in the E-W direction (along the x-axis)
\item lag.y: lag in the S-N direction (along the y-axis)
\item n: number of differences/vectors the estimation is based on
\item variogram: estimated variogram value
\item estimator: used estimator}
}
\description{
Returns robust and non-robust directional variogram estimations in the four main directions south-north, east-west, southeast-northwest and southwest-northeast for data on a two-dimensional regular grid such as image data.
}
\details{
The variogram is a measure for the spatial dependency. For a given spatial lag vector \eqn{\boldsymbol{h}} and an intrinsic stationary and mean stationary random field \eqn{\{Z(\boldsymbol{s}), \boldsymbol{s}\in I \}} it is defined as
\deqn{2 \gamma(\boldsymbol{h}) = \text{Var}(Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h})),~\boldsymbol{s},\boldsymbol{s}+\boldsymbol{h}\in I.}
Three different estimators can be used here to
estimate the variogram for a given two-dimensional lag \eqn{\boldsymbol{h}} vector in one of the four directions south-north, east-west, southeast-northwest and southwest-northeast for data on a two-dimensional regular grid.
For the different directions the lag vectors has the following forms:
\itemize{
\item south-north: \eqn{\boldsymbol{h} = \left(\begin{array}{c} 0 \\ x \end{array}\right)}
\item east-west: \eqn{\boldsymbol{h} = \left(\begin{array}{c} x \\ 0 \end{array}\right)}
\item southwest-northeast: \eqn{\boldsymbol{h} = \left(\begin{array}{c} x \\ x \end{array}\right)}
\item southeast-northwest: \eqn{\boldsymbol{h} = \left(\begin{array}{c} x \\ -x \end{array}\right)}.
}
If \code{estimator = "matheron"} the popular Matheron variogram estimator is used (Matheron, 1962):
\deqn{2\widehat{\gamma}(\boldsymbol{h}) = \frac{1}{|N(\boldsymbol{h})|} \sum_{(\boldsymbol{s}_i, \boldsymbol{s}_j)\in N(\boldsymbol{h})} (Z(\boldsymbol{s}_i) - Z(\boldsymbol{s}_j))^2 }
with \eqn{N(\boldsymbol{h}) = \{(\boldsymbol{s}_i, \boldsymbol{s}_j)\in I^2: \boldsymbol{s}_j - \boldsymbol{s}_i = \boldsymbol{h}\}} being the set of all pairs in distance \eqn{h}.
This estimator is non robust (Genton, 1998, Gierse & Fried 2025).

A robust alternative especially for random fields with isolated outliers is the Genton variogram estimator proposed by Genton (1998) (\code{estimator = "genton"}).
\eqn{V_i(\boldsymbol{h})} is the \eqn{i}-th element of the set \eqn{V(\boldsymbol{h}) = \{Z(\boldsymbol{s}_l) - Z(\boldsymbol{s}_m): (\boldsymbol{s}_l, \boldsymbol{s}_m)\in N(\boldsymbol{h})\}} containing all differences for locations with distance \eqn{\boldsymbol{h}}.
The estimator is given by:
\deqn{2\widehat{\gamma}(\boldsymbol{h}) = \left(c \cdot \left(|V_i(\boldsymbol{h}) - V_j(\boldsymbol{h})|: i<j\right)_{(k)}\right)^2}
with \eqn{k = \begin{pmatrix}\left[\frac{|N(\boldsymbol{h})|}{2}\right] + 1 \\ 2\end{pmatrix}},  \eqn{[a]} denoting the integer part of \eqn{a}, \eqn{c} being a consistency factor,
which is approximately equal to \eqn{2.22} in large Gaussian samples, and \eqn{(\cdot)_{(k)}} denoting the \eqn{k}-th order statistic.

Gierse & Fried (2025) proposed two estimators for random fields with an outlier block. Both are multivariate variogram estimators, which estimate the variogram for a set of lags.
 The option \code{estimator = "mcd.org"} uses the proposed MCD.org estimator. For a set of lag vectors \eqn{\boldsymbol{h}_1, \ldots, \boldsymbol{h}_{h_\text{max}}} vectors of the following
 structure are built
 \deqn{\boldsymbol{V} = \begin{pmatrix} Z(\boldsymbol{s}) \\ Z(\boldsymbol{s}+\boldsymbol{h}_1) \\ \vdots \\ Z(\boldsymbol{s}+\boldsymbol{h}_{h_{\max}})
 \end{pmatrix}}
 The (reweighted) multivariate covariance determinant estimator, a multivariate and highly robust covariance estimator, is used to robust estimate the variance-covariance matrix of the vectors. The function \code{covMcd} of the package \pkg{robustbase}
 is used for estimation. The covariance matrix has the following structure
 \deqn{	\boldsymbol{\Sigma}_{\boldsymbol{V}} = \begin{pmatrix}
 a_0 & a_1 & a_2 & \ldots & a_{h_{\max}} \\
 a_1 & a_0 & \ddots & \ddots & \vdots \\
 a_2 & \ddots & \ddots & \ddots & a_2    \\
 \vdots & \ddots & \ddots & \ddots & a_1 \\
 a_{h_{\max}} & \ldots & a_2 & a_1 & a_0
 \end{pmatrix}
 \in \mathbb{R}^{(h_{\max} + 1) \times (h_{\max}+1)}}
 with \eqn{a_0 = \text{Var}(Z(\boldsymbol{s})), a_1 = \text{Cov}(Z(\boldsymbol{s}), Z(\boldsymbol{s} +\boldsymbol{h}_1)), a_2 = \text{Cov}(Z(\boldsymbol{s}), Z(\boldsymbol{s}+\boldsymbol{h}_2))} and so on.
 The different estimates for the variance and covariances are averaged. The MCD.org variogram estimator is calculated by
 \eqn{2\widehat{\gamma}(\boldsymbol{h}) = 2[\widehat{\text{Var}}(Z(\boldsymbol{s})) - \widehat{\text{Cov}}(Z(\boldsymbol{s}), Z(\boldsymbol{s}+\boldsymbol{h}))]}.

 The option \code{estimator = "mcd.diff"} performs the proposed MCD.diff estimator. For this estimator the following vectors of differences of the process are built
 \deqn{\boldsymbol{W} = \begin{pmatrix}
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_1) \\
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_2) \\
 \vdots \\
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_{h_{\max}})
 \end{pmatrix}}
 Again the function \code{covMcd} of the package \pkg{robustbase} to robust estimate the variance-covariance matrix of these vectors. The estimated
 variogram values are then obtained from the diagonal of the estimated covariance matrix
 \deqn{2\widehat{\gamma}(\boldsymbol{h}) = \text{diag}\left(\widehat{\boldsymbol{\Sigma}}\right)}

MCD.diff as well as MCD.org uses the MCD estimator for the robust estimation of the covariance matrix of the vectors.
These estimator requires at least twice as many vectors as the dimension of the vectors.
Therefore \deqn{h_{\max}} must  not be too large relative to the grid size, so that enough vectors can be generated.
}
\examples{
\donttest{

## Simulate an isotropic gaussian random field without outliers
dat <- simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6), n.it = 1)

dat <- cbind(dat$data, dat$grid)

## calculate the variogram using MCD.diff estimator for all four directions
varog <- variogram_est(data = dat, hmax = c(7,7,5,5),
                       direction = c("S-N", "E-W", "SW-NE", "SE-NW"),
                       estimator = "mcd.diff")
}


}
\references{
Genton, M. (1998). Highly robust variogram estimation. \emph{Mathematical Geology}, 30, 213-221. \doi{https://doi.org/10.1023/A:1021728614555}

Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}

Matheron, G. (1962). Traité de géostatistique appliquée, Tome I. \emph{Mémoires du Bureau de Recherches Géologiques et Miniéres}, no. 14, Editions Technip, 333
}
\seealso{
\code{\link{variogram_est_general}}
}
