% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_est_general.R
\name{variogram_est_general}
\alias{variogram_est_general}
\title{Nonparametric directional variogram estimation}
\usage{
variogram_est_general(
  data,
  h,
  estimator = "all",
  cp = NULL,
  reweighting = TRUE,
  ...
)
}
\arguments{
\item{data}{A dataframe with three columns containing the data (first column) and the grid (second and third column).}

\item{h}{A matrix with two columns containing the lag vectors for which the variogram should be estimated. Each row contain one lag vector.}

\item{estimator}{A character string or vector of character strings specifying the estimators to be used. Possible estimators are \code{"matheron"}, \code{"genton"}, \code{"mcd.org"} or \code{"mcd.diff"}. If \code{"all"} all four estimators are used. See details; Default is "all".}

\item{cp}{A vector containing possible finite sample correction factors for each direction for the MCD based variogram estimators. If it has length 1 the same correction factor is used for all directions.}

\item{reweighting}{Logical indicating if the reweighted version of the MCD based variogram estimators should be used; default is \code{TRUE}.}

\item{...}{Optional argument for the function \code{covMcd} of the package \pkg{robustbase} used for the MCD based variogram estimators}
}
\value{
An object of class "varioRobgen" which is a list with a dataframe for each direction with columns
\itemize{
\item lag.x: lag in the E-W direction (along the x-axis)
\item lag.y: lag in the S-N direction (along the y-axis)
\item n: number of differences/vectors the estimation is based on
\item variogram: estimated variogram value
\item estimator: used estimator}
}
\description{
Returns robust and non-robust directional variogram estimations in arbitrary directions for data on a two-dimensional regular grid.
}
\details{
The variogram characterizes the spatial dependence of a random field. For a given lag vector \eqn{\boldsymbol{h}} and an intrinsically and mean stationary random field \eqn{\{Z(\boldsymbol{s}), \boldsymbol{s}\in I \}} it is defined as
\deqn{2 \gamma(\boldsymbol{h}) = \text{Var}(Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h})),~\boldsymbol{s},\boldsymbol{s}+\boldsymbol{h}\in I.}
Three different estimators are implemented here to
estimate the variogram for arbitrary given two-dimensional lag vectors \eqn{h} for data on a two-dimensional regular grid.

If \code{estimator = "matheron"} the popular Matheron variogram estimator is used (Matheron, 1962):
\deqn{2\widehat{\gamma}(\boldsymbol{h}) = \frac{1}{|N(\boldsymbol{h})|} \sum_{(\boldsymbol{s}_i, \boldsymbol{s}_j)\in N(\boldsymbol{h})} (Z(\boldsymbol{s}_i) - Z(\boldsymbol{s}_j))^2 }
with \eqn{N(\boldsymbol{h}) = \{(\boldsymbol{s}_i, \boldsymbol{s}_j)\in I^2: \boldsymbol{s}_j - \boldsymbol{s}_i = \boldsymbol{h}\}} being the set of all pairs in distance \eqn{h}.
This estimator is non robust (Genton, 1998, Gierse & Fried 2025).

A robust alternative, particularly suitable for random fields containing isolated outliers, is the Genton variogram estimator proposed by Genton (1998) (\code{estimator = "genton"}).
Let \eqn{V_i(\boldsymbol{h})} denote the \eqn{i}-th element of the set \eqn{V(\boldsymbol{h}) = \{Z(\boldsymbol{s}_l) - Z(\boldsymbol{s}_m): (\boldsymbol{s}_l, \boldsymbol{s}_m)\in N(\boldsymbol{h})\}}
which contains all pairwise differences of observations separated by the lag vector \eqn{\boldsymbol{h}}.
The estimator is defined as:
\deqn{2\widehat{\gamma}(\boldsymbol{h}) = \left(c \cdot \left(|V_i(\boldsymbol{h}) - V_j(\boldsymbol{h})|: i<j\right)_{(k)}\right)^2}
where \eqn{k = \begin{pmatrix}\left[\frac{|N(\boldsymbol{h})|}{2}\right] + 1 \\ 2\end{pmatrix}},  \eqn{[a]} denoting the integer part of \eqn{a}, \eqn{c} is a consistency factor
(approximately equal to \eqn{2.22} in large Gaussian samples) and \eqn{(\cdot)_{(k)}} denoting the \eqn{k}-th order statistic.

Gierse & Fried (2025) proposed two estimators designed for random fields affected by blocks of outliers. Both approaches are multivariate variogram estimators that simultaneously estimate the variogram over a set of lags.
 The option \code{estimator = "mcd.org"} uses the proposed MCD.org estimator. For a collection of lag vectors \eqn{\boldsymbol{h}_1, \ldots, \boldsymbol{h}_{h_\text{max}}}, vectors of the following
 \deqn{\boldsymbol{V} = \begin{pmatrix} Z(\boldsymbol{s}) \\ Z(\boldsymbol{s}+\boldsymbol{h}_1) \\ \vdots \\ Z(\boldsymbol{s}+\boldsymbol{h}_{h_{\max}})
 \end{pmatrix}}
 are constructed.
 To robustly estimate the associated variance–covariance matrix, the (reweighted) minimum covariance determinant (MCD) estimator is employed.
 This estimator is a highly robust multivariate method for covariance estimation. The function \code{covMcd} from the robustbase package is used for this.
 The resulting covariance matrix has the structure
 \deqn{	\boldsymbol{\Sigma}_{\boldsymbol{V}} = \begin{pmatrix}
 a_0 & a_1 & a_2 & \ldots & a_{h_{\max}} \\
 a_1 & a_0 & \ddots & \ddots & \vdots \\
 a_2 & \ddots & \ddots & \ddots & a_2    \\
 \vdots & \ddots & \ddots & \ddots & a_1 \\
 a_{h_{\max}} & \ldots & a_2 & a_1 & a_0
 \end{pmatrix}
 \in \mathbb{R}^{(h_{\max} + 1) \times (h_{\max}+1)}}
 where \eqn{a_0 = \text{Var}(Z(\boldsymbol{s})), a_1 = \text{Cov}(Z(\boldsymbol{s}), Z(\boldsymbol{s} +\boldsymbol{h}_1)), a_2 = \text{Cov}(Z(\boldsymbol{s}), Z(\boldsymbol{s}+\boldsymbol{h}_2))} and so forth.
 The variance and covariance estimates obtained from this matrix are averaged accordingly. The MCD.org variogram estimator is then given by
 \eqn{2\widehat{\gamma}(\boldsymbol{h}) = 2[\widehat{\text{Var}}(Z(\boldsymbol{s})) - \widehat{\text{Cov}}(Z(\boldsymbol{s}), Z(\boldsymbol{s}+\boldsymbol{h}))]}.

 The option \code{estimator = "mcd.diff"} implements the proposed MCD.diff estimator. For this approach, vectors of pairwise differences of the process are constructed as
 \deqn{\boldsymbol{W} = \begin{pmatrix}
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_1) \\
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_2) \\
 \vdots \\
 Z(\boldsymbol{s}) - Z(\boldsymbol{s}+\boldsymbol{h}_{h_{\max}})
 \end{pmatrix}}
 As before, the function \code{covMcd} from the package \pkg{robustbase} is used to obtain a robust estimate of the variance–covariance matrix of these vectors.
 The variogram estimates are then derived from the diagonal elements of the estimated covariance matrix, i.e.,
 \deqn{2\widehat{\gamma}(\boldsymbol{h}) = \text{diag}\left(\widehat{\boldsymbol{\Sigma}}\right)}

MCD.diff as well as MCD.org uses the MCD estimator for the robust estimation of the covariance matrix of the vectors.
These estimator requires at least twice as many vectors as the dimension of the vectors.
Therefore \eqn{h_{\max}} must not be too large relative to the grid size, so that enough vectors can be generated.

Both the MCD.diff and MCD.org estimators rely on the MCD approach for the robust estimation of the covariance matrix of the constructed vectors.
These estimators require that the number of available vectors is at least twice the dimension of the vectors. Consequently, the choice of
\eqn{h_{\max}} should not be too large relative to the grid size, ensuring that a sufficient number of vectors can be formed.

If the set of lag vectors of interest contains only vectors  in the following directions south-north, east-west, southeast-northwest and southwest-northeast,
then the function \code{variogram_est} should be used, as it has better runtime performance.
}
\examples{
\donttest{

## Simulate an anisotropic gaussian random field without outliers
dat <- simulate_grf(gridsize = c(20, 20), param.variogram = c(1, 6),
                    aniso.param = c(pi/4, 2), n.it = 1)

dat <- cbind(dat$data, dat$grid)

## calculate the variogram using all implemented estimator for different lags
varog <- variogram_est_general(data = dat,
                               h = rbind(c(2,1), c(4,2), c(1,2), c(2,4)),
                               estimator = "all")
}

}
\references{
Genton, M. (1998). Highly robust variogram estimation. \emph{Mathematical Geology}, 30, 213-221. \doi{https://doi.org/10.1023/A:1021728614555}

Gierse, J., & Fried, R. (2025). Nonparametric directional variogram estimation in the presence of outlier blocks. \emph{Statistical Papers}, 66(134). \doi{https://doi.org/10.1007/s00362-025-01754-2}

Matheron, G. (1962). Traité de géostatistique appliquée, Tome I. \emph{Mémoires du Bureau de Recherches Géologiques et Miniéres}, no. 14, Editions Technip, 333
}
\seealso{
\code{\link{variogram_est}}
}
